requirements_drl.txt
# DRL Training Dependencies
stable-baselines3==2.2.1
torch>=1.9.0
tensorboard>=2.8.0
matplotlib>=3.5.0
seaborn>=0.11.0
pandas>=1.3.0
numpy>=1.21.0
pyyaml>=6.0
gymnasium>=0.27.0
